projectx/requirements.txt
apturl==0.5.2
argcomplete==1.8.1
asn1crypto==0.24.0
astroid==2.2.5
beautifulsoup4==4.6.0
boto3==1.9.230
botocore==1.12.230
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.2.5
django-storages==1.7.2
djangorestframework==3.10.3
docutils==0.15.2
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
isort==4.3.21
jmespath==0.9.4
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.2
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.4.4
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
progressbar==2.3
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
python-gnupg==0.4.1
pytz==2019.2
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
s3transfer==0.2.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
software-center-aptd-plugins==0.0.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
Ubuntu-Make==16.11.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.3
usb-creator==0.3.3
vboxapi==1.0
wadllib==1.3.2
webencodings==0.5
wrapt==1.11.2
xkit==0.0.0
zope.interface==4.3.2
